app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
colorama==0.4.6
    # via loguru
loguru==0.7.3
    # via image-host (pyproject.toml)
psycopg==3.3.2
    # via image-host (pyproject.toml)
psycopg2-binary==2.9.11
    # via image-host (pyproject.toml)
python-dotenv==1.2.1
    # via image-host (pyproject.toml)
tzdata==2025.3
    # via psycopg
win32-setctime==1.2.0
    # via loguru
